Renamed `moq_input` to `moq_server_input`
