+++ smelter-core/src/pipeline/moq/mod.rs
@@ -1,7 +1,7 @@
 mod connection;
-mod moq_input;
+mod moq_server_input;
 mod server;
 mod state;
 
-pub use moq_input::MoqServerInput;
+pub use moq_server_input::MoqServerInput;
 pub(super) use server::{MoqPipelineState, MoqServer, spawn_moq_server};
